requirements.txt
# TransWeave-Finance: Reproduction of Goertzel (2025)
# Python 3.10+, PyTorch 2.x. Install: pip install -r requirements.txt

# Data pipeline (Stage 1)
ccxt>=4.0
dune-client>=1.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0
pyarrow>=14.0

# Core ML
torch>=2.0,<3.0
scipy>=1.10
pyyaml>=6.0

# Jump-diffusion + TFT (Stage 2-3)
arch>=7.0
hmmlearn>=0.3
pot>=0.9

# Evaluation
properscoring>=0.1

# Plotting
matplotlib>=3.7
seaborn>=0.12

# Utilities
tqdm>=4.65
tabulate>=0.9
